move: firebase -> device
