+++ src/main/java/com/nexters/goalpanzi/application/member/MemberService.java
@@ -8,8 +8,8 @@ import com.nexters.goalpanzi.application.member.event.DeleteMemberEvent;
 import com.nexters.goalpanzi.application.member.event.UpdateDeviceTokenEvent;
 import com.nexters.goalpanzi.application.member.event.UpdatePushActivationStatusEvent;
 import com.nexters.goalpanzi.domain.auth.repository.RefreshTokenRepository;
-import com.nexters.goalpanzi.domain.firebase.Device;
-import com.nexters.goalpanzi.domain.firebase.repository.DeviceRepository;
+import com.nexters.goalpanzi.domain.device.Device;
+import com.nexters.goalpanzi.domain.device.repository.DeviceRepository;
 import com.nexters.goalpanzi.domain.member.Member;
 import com.nexters.goalpanzi.domain.member.repository.MemberRepository;
 import com.nexters.goalpanzi.exception.AlreadyExistsException;
