+++ src/main/java/com/nexters/goalpanzi/domain/device/Devices.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.domain.firebase;
+package com.nexters.goalpanzi.domain.device;
 
 import lombok.RequiredArgsConstructor;
